JAJA NO SE QUE HICE PERO LO ARREGLE

+++ app/src/main/java/pe/edu/ulima/pm/pokeapp/model/PokemonManager.kt
@@ -48,7 +48,7 @@ class PokemonManager(context: Context) {
         -------------------------------------------------------------------------------------------*/
     }
 
-    fun getPokemon(i:Int): Pokemon{
+    fun getPokemon(i:Long): Pokemon{
         var pk : Pokemon? = null
         dbFirebase.collection("pokemon")
             .get()
